# openapi.yaml
openapi: 3.0.0
info:
  title: Wheres Poppin API
  version: 1.0.0
  description: API for Wheres Poppin.
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://wherespoppin-server.onrender.com/
    description: Prodcution server
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: string
          description: Unique ID of the user
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        phone:
          type: string
          description: User's phone number
        gender:
          type: string
          enum: ["male", "female", "other"]
          description: User's gender
        dob:
          type: string
          format: date
          description: Date of birth of the user
        profileImage:
          $ref: '#/components/schemas/Media'
          description: Profile image of the user
...
    Media:
      type: object
      properties:
        url:
          type: string
          description: URL of the media file
        type:
          type: string
          description: Type of the media (image, video, etc.)
        size:
          type: integer
          description: Size of the media in bytes
      example:
        url: "https://example.com/media/event1.jpg"
        type: "image"
...
paths:
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - Users
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags:
        - Users
      summary: Update a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
    delete:
      tags:
        - Users
      summary: Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
...